ML-39 Move permissions file
